레이저 찍먹 구현

+++ Space1945/Assets/Prefabs/Resources/Enemy/Patterns/BulletType/NormalEnemy.cs
@@ -7,7 +7,7 @@ public class NormalEnemy : MonoBehaviour
     // 그냥 아무런 효과가 없는 탄환
     // 충돌시 데미지
     float speed;
-    int crash_damage;
+    float crash_damage;
 
     void Start()
     {
