+++ Space1945/Assets/Scripts/Parts/Attack/BulletType/Explosion.cs
@@ -6,14 +6,13 @@ public class Explosion : MonoBehaviour
 {
     //폭발형 탄환
     //충돌시 일정 범위 내의 적에게 데미지
-    //해당 탄환은 탄두로 적에게 데미지를 입힐 수 없음 Mob info의 충돌보다 아래의 충돌이 더 빠르기 때문에
-    //따라서 탄두 자체의 crash_damage로는 적에게 데미지를 입힐 수 없으므로 Boom클래스 내의 폭파데미지로 적에게 데미지를 입혀야함
+    // 탄두 자체로도 소량의 데미지 부여 가능
     public GameObject boom;
 
     float shot_angle;
     Vector2 normalized_angle;
     float speed;
-    int crash_damage;
+    float crash_damage;
 
     void Start()
     {
